feat: remove test for invalid point or vector.
Instead of worrying whether a point or vector
is valid, I'm choosing to continue calculations
as implemented and handle invalid point or
vector as the usage sites.

If there are too much duplication at the
usage sites, then it would be a good time
to consider having Tuple validate its subtypes.